requirements.txt
dj-database-url==0.5.0
Django==2.2.1
djangorestframework==3.9.4
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
